compose: Move tmpfiles generation into tmpfiles module
Prep for further work around postprocessing; just makes
sense to have the tmpfiles code in the same place.

No functional changes.
